Make house selection event-driven
I feel as though this solution is overkill for the task at hand. It just
added a lot of new code and complexity. However, it also removed
unnecessary assignments and computations from the render loop. I think
that, in the long run, this strategy is how OpenFL intends for games to
be implemented.